Fixes for warnings (redundant casts, raw types).
git-svn-id: svn+ssh://src.earth.threerings.net/clyde/trunk@25 edc67d0a-ffea-dc11-983e-003048321218

+++ src/java/com/threerings/math/FloatMath.java
@@ -169,7 +169,7 @@ public class FloatMath
      */
     public static float abs (float v)
     {
-        return (float)Math.abs(v);
+        return Math.abs(v);
     }
 
     /**
